posix: tty: remove unnecessary u32 cast
This fixes the following clippy warning:

warning: casting to the same type is unnecessary (`u32` -> `u32`)
   --> src/posix/tty.rs:461:12
    |
461 |         Ok(termios2.c_ospeed as u32)
    |            ^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `termios2.c_ospeed`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
    = note: `#[warn(clippy::unnecessary_cast)]` on by default

+++ src/posix/tty.rs
@@ -458,7 +458,7 @@ impl SerialPort for TTYPort {
 
         assert!(termios2.c_ospeed == termios2.c_ispeed);
 
-        Ok(termios2.c_ospeed as u32)
+        Ok(termios2.c_ospeed)
     }
 
     /// Returns the port's baud rate
